+++ test/test_grouped_paulis.py
@@ -20,7 +20,7 @@ import unittest
 from qiskit.quantum_info import pauli_group
 
 from test.common import QiskitAquaTestCase
-from qiskit_aqua.operator import Operator
+from qiskit_aqua import Operator
 
 
 class TestGroupedPaulis(QiskitAquaTestCase):
